Fix zh_cn.json and a small request
This a fix of zh_cn.json for age of exile . And I'm wondering that if I can repost Age of Exile to a Chinese forum , you know , because of GFW , it's hard for we Chinese players to know some good mods.This is an excellent mod,and I hope more Chinese player will know it. I will link to curseforge pages and will keep everything you have asked.